pre-populate the text field
