--- scp/android/app/src/google-services.json
@@ -1,80 +1,0 @@
-{
-  "project_info": {
-    "project_number": "56350565580",
-    "firebase_url": "https://project-avocado-8b3e1.firebaseio.com",
-    "project_id": "project-avocado-8b3e1",
-    "storage_bucket": "project-avocado-8b3e1.appspot.com"
-  },
-  "client": [
-    {
-      "client_info": {
-        "mobilesdk_app_id": "1:56350565580:android:8547c8a5bd5b42ad",
-        "android_client_info": {
-          "package_name": "in.ac.nitrkl.scp.scp"
-        }
-      },
-      "oauth_client": [
-        {
-          "client_id": "56350565580-gg1kct8o1bmt01cll5gah3sq8ed9e8v2.apps.googleusercontent.com",
-          "client_type": 1,
-          "android_info": {
-            "package_name": "in.ac.nitrkl.scp.scp",
-            "certificate_hash": "d798d5db0bac76e880651da66dd76a2c851b3870"
-          }
-        },
-        {
-          "client_id": "56350565580-ctmefvbl38c2asej3oob80qpa98gt5r0.apps.googleusercontent.com",
-          "client_type": 1,
-          "android_info": {
-            "package_name": "in.ac.nitrkl.scp.scp",
-            "certificate_hash": "fede7a2adb7cd9ff3c99e67c3747c5570859cfee"
-          }
-        },
-        {
-          "client_id": "56350565580-utd10uablbet6n87b7c14pftiks033ph.apps.googleusercontent.com",
-          "client_type": 1,
-          "android_info": {
-            "package_name": "in.ac.nitrkl.scp.scp",
-            "certificate_hash": "9903728f2b22e555eb699cc7ba733f70c23c5551"
-          }
-        },
-        {
-          "client_id": "56350565580-crjjbo1pk6cv4p3g0a9ruutr2nghc2qc.apps.googleusercontent.com",
-          "client_type": 1,
-          "android_info": {
-            "package_name": "in.ac.nitrkl.scp.scp",
-            "certificate_hash": "c969fd52b5b752b5dc32fdf17730c7b5c6c51cad"
-          }
-        },
-        {
-          "client_id": "56350565580-i34vpsajlmifo1kvvdtt9j1uu4v6iqt8.apps.googleusercontent.com",
-          "client_type": 1,
-          "android_info": {
-            "package_name": "in.ac.nitrkl.scp.scp",
-            "certificate_hash": "c483a090db6a3720bbcfe126ec167cf61ccac6c6"
-          }
-        },
-        {
-          "client_id": "56350565580-ca3cjp9u6f7qmdrbh7234ars08ni55cf.apps.googleusercontent.com",
-          "client_type": 3
-        }
-      ],
-      "api_key": [
-        {
-          "current_key": "AIzaSyAe9nhpuDDVj4p7TD409IVmbwN4-dVEaDI"
-        }
-      ],
-      "services": {
-        "appinvite_service": {
-          "other_platform_oauth_client": [
-            {
-              "client_id": "56350565580-ca3cjp9u6f7qmdrbh7234ars08ni55cf.apps.googleusercontent.com",
-              "client_type": 3
-            }
-          ]
-        }
-      }
-    }
-  ],
-  "configuration_version": "1"
-}
